MapsActivity: Trigger WebView back for back key
There’s an issue with this implementation: the user needs to exhaust WebView’s history/stack before they can exit out of the app (because canGoBack will still return true until the history reaches 0). This isn’t much of a problem per se if the user don’t dart left and right, but let’s not assume that way (and maybe try finding a fix!).